Created a handlers package to make main.go more readable, and also added a frontEnd directory for client-related files
